Add stdlib stub and split InMemoryLedger to its own file
